+++ R/layouts.R
@@ -105,6 +105,7 @@ time_ordered_coords <- function(
         fixed_time = fixed_time,
         adjust_exposure_outcome = adjust_exposure_outcome,
         force_y = force_y,
+        node_scale = ggdag_option("node_size", 16) / 16,
         ...
       )
     }